For #1434 - Remove symptom checker, add symptom poster skeleton
