+++ client/flutter/lib/pages/main_pages/routes.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/cupertino.dart';
 import 'package:provider/provider.dart';
-import 'package:who_app/api/content/content_loading.dart';
 import 'package:who_app/api/content/content_store.dart';
 import 'package:who_app/api/stats_store.dart';
 import 'package:who_app/api/user_preferences_store.dart';
@@ -9,15 +8,12 @@ import 'package:who_app/generated/l10n.dart';
 import 'package:who_app/pages/about_page.dart';
 import 'package:who_app/pages/facts_carousel_page.dart';
 import 'package:who_app/pages/main_pages/app_tab_router.dart';
-import 'package:who_app/pages/main_pages/check_up_intro_page.dart';
+import 'package:who_app/pages/main_pages/check_up_poster_page.dart';
 import 'package:who_app/pages/main_pages/recent_numbers.dart';
 import 'package:who_app/pages/news_page.dart';
 import 'package:who_app/pages/onboarding/onboarding_page.dart';
 import 'package:who_app/pages/protect_yourself.dart';
 import 'package:who_app/pages/question_index.dart';
-import 'package:who_app/pages/symptom_checker/symptom_checker_model.dart';
-import 'package:who_app/pages/symptom_checker/symptom_checker_results_page.dart';
-import 'package:who_app/pages/symptom_checker/symptom_checker_view.dart';
 import 'package:who_app/pages/travel_advice.dart';
 
 class Routes {
@@ -39,12 +35,9 @@ class Routes {
           dataSource: Provider.of<ContentStore>(context),
           title: S.of(context).homePagePageButtonQuestions,
         ),
-    '/symptom-checker': (context) => CheckUpIntroPage(),
-    '/symptom-checker-survey': (context) => SymptomCheckerView(
-        contentService: Provider.of<ContentService>(context)),
-    '/symptom-checker-results': (context) => SymptomCheckerResultsPage(
-        model:
-            ModalRoute.of(context).settings.arguments as SymptomCheckerModel),
+    '/symptom-checker': (context) => CheckUpPosterPage(
+          dataSource: Provider.of<ContentStore>(context),
+        ),
     '/news': (context) => NewsIndexPage(
           dataSource: Provider.of<ContentStore>(context),
         ),
